% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffExprAnalysis.R
\name{calculateDiffExprEdgeR}
\alias{calculateDiffExprEdgeR}
\title{Differential Expression Analysis with edgeR}
\usage{
calculateDiffExprEdgeR(y, experimental_info = NULL, gene_info = NULL,
  group = NULL, reference = NULL, filter = T, filter.expr.th = 1,
  filter.sample.th = 2, normalize.using = "cpm", tlen = NULL,
  method = "exact", anovalike = F, design = NULL, cf = NULL,
  contrast = NULL, return.y = F)
}
\arguments{
\item{y}{count matrix (or data.frame) or object of class DGEList}

\item{experimental_info}{data.frame containing informations about the experimental conditions for each sample}

\item{group}{conditions to test}

\item{method}{statistcal test to be performed (exact, lrt, qlf)}
}
\value{
results of the test (data.frame)
}
\description{
Implementation of statistical models and tests 
for differential expresion analysis using edgeR
software package
}
\examples{
\dontrun{
    dea <- calculateDiffExprEdgeR(y, expreimental_info
    , group = experimental_info$group)
}


}
